requirements.txt
torch==2.0.1
torchvision==0.15.2
flask==2.3.2
pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.74
scikit-learn==1.3.0
matplotlib==3.7.2
tqdm==4.65.0